[#8] Remove MonitoringGroupListActivity
The file was created when I did the skeleton code for groups at the beginning of release 1. However, we then agreed that each user can only have 1 monitoring group in release 1. Hence, the app should bring the users straight to the monitoring group page instead of showing a list of monitoring groups which is either 0 or 1 group.